User controllers edited

+++ src/controllers/user.controllers.ts
@@ -5,10 +5,18 @@ export const getAllUsers = async (req: Request, res: Response) => {
   try {
     const allUsers = await prisma.user.findMany({
       include:{
-        movies: true
+        movies: {
+          include: {
+            genre: true
+          }
+        }
       }
     });
-    res.status(200).send(allUsers);
+    res.status(200).send({
+      msg: "All users",
+      data: allUsers,
+      type: "array"
+    });
   } catch (error) {
     res.status(400).send(error);
   }
@@ -16,11 +24,22 @@ export const getAllUsers = async (req: Request, res: Response) => {
 
 export const createUser = async (req: Request, res: Response) => {
   const { name, email, password } = req.body;
+
+  if(!name || !email || !password) {
+    return res.status(400).send({
+      message: "The fields name, email and password are required"
+    })
+  }
+  
   try {
     const newUser = await prisma.user.create({
       data: {name, email, password}
     });
-    res.status(201).send(newUser);
+    res.status(201).send({
+      msg: "User created successfully",
+      data: newUser,
+      type: typeof newUser
+    });
   } catch (error) {
     res.status(400).send(error);
   }
@@ -30,12 +49,22 @@ export const updateUser = async(req: Request, res: Response) => {
     const {name, email, password}= req.body
     const {userID} = req.params
 
+    if(!name || !email || !password) {
+      return res.status(400).send({
+        message: "The fields name, email and password are required"
+      })
+    }
+
     try{
         const userUpdated = await prisma.user.update({
           where:{id:userID},
           data:{name, email, password}
         })
-        res.status(201).send(userUpdated)
+        res.status(201).send({
+          msg: "User updated successfully",
+          data: userUpdated,
+          type: typeof userUpdated
+        })
     }catch(error){
         res.status(400).send(error)
     }
@@ -43,11 +72,64 @@ export const updateUser = async(req: Request, res: Response) => {
 
 export const deleteUser = async(req: Request, res: Response) => {
   const {userID} = req.params
+
+  if(!userID){
+    return res.status(400).send({
+      message: "The field userID is required"
+    })
+  }
+
     try{
-        const userDeleted = await prisma.user.delete({
-          where:{id:userID}
+        const user = await prisma.user.findUnique({
+          where: {
+            id: userID,
+          },
+          include: {
+            movies:{
+              include: {
+                genre: true,
+              }
+            }
+          }
         })
-        res.status(201).send(userDeleted)
+
+        if(!user){
+          return res.status(404).send({
+            message: "User not found"
+          })
+        }
+
+        const genreOnMoviesToDelete = user.movies.flatMap((movie) =>
+          movie.genre.map((genre) => ({
+            movieID: movie.id,
+            genreID: genre.id
+          }))
+        )
+
+        await prisma.movieGenre.deleteMany({
+          where: {
+            OR: genreOnMoviesToDelete,
+          }
+        })
+
+        const deletedMovies = await prisma.movie.deleteMany({
+          where: {
+            userID: userID
+          }
+        })
+
+        const deletedUser = await prisma.user.delete({
+          where: {
+            id: userID,
+          }
+        })
+        res.status(200).send({
+          message: "User deleted successfully",
+          data: {
+            deletedMovies,
+            deletedUser
+          }
+        })
     }catch(error){
         res.status(400).send(error)
     }
